requirements.txt
fastapi
fastapi[standard]
requests
psycopg2-binary